pra 65 feat: create style for ErrorBoundary, update BlockComments (#19)
