+++ packages/client/src/pages/Forum/components/CommentsPage/CommentsPage.module.scss
@@ -1,0 +1,11 @@
+@import 'src/index.module.scss';
+
+.block {
+  display: flex;
+  flex-flow: column;
+  height: 500px;
+}
+
+.button_new_post {
+  margin: 10px 50px 40px;
+}
