+++ packages/client/src/pages/Forum/components/CommentsPage/index.ts
@@ -1,0 +1,1 @@
+export { CommentsPage } from './CommentsPage';
