+++ packages/client/src/pages/Forum/components/CommentsPage/types.ts
@@ -1,0 +1,5 @@
+import { ITopic } from "../../part/Topic/types";
+
+export interface Props {
+  topic: ITopic;
+}
