+++ packages/client/src/pages/Forum/part/BlockComments/BlockComment.module.scss
@@ -1,14 +1,5 @@
 @import 'src/index.module.scss';
 
-.block {
-  display: flex;
-  flex-flow: column;
-  height: 500px;
-}
-
-.button_new_post {
-  margin: 10px 50px 40px;
-}
 
 .block_topics {
   display: flex;
