+++ packages/client/src/pages/GameScreen/GameScreen.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import './GameScreen.module.scss';
 
-import {GameComponent} from '../../components/Game';
+import { GameComponent } from '../../components/Game';
 import { BackgroundLayout } from 'src/layouts/BackgroundLayout';
 export const GameScreen = () => {
-    return ( 
+    return (
         <BackgroundLayout>
-             <GameComponent/>
+            <GameComponent />
         </BackgroundLayout>
     );
 };
