+++ packages/client/src/pages/Leaderboard/components/LeaderboardLine.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { ErrorBoundary } from 'src/components/ErrorBoundary';
 import styles from './LeaderboardLine.module.scss';
 
 interface ITeam {
@@ -14,12 +15,14 @@ interface Props {
 export const LeaderboardLine = ({ idx, team }: Props) => {
   const { name, score, players } = team;
   return (
-    <div className={styles.leader_line}>
-      <span className={styles.leader_line_item}>{name}</span>
-      <span className={styles.leader_line_item}>{idx + 1}</span>
-      <span className={styles.leader_line_item}>{score}</span>
-      <span className={styles.leader_line_item}>{players[0]}, </span>
-      <span className={styles.leader_line_item}>{players[1]}</span>
-    </div>
+    <ErrorBoundary>
+      <div className={styles.leader_line}>
+        <span className={styles.leader_line_item}>{name}</span>
+        <span className={styles.leader_line_item}>{idx + 1}</span>
+        <span className={styles.leader_line_item}>{score}</span>
+        <span className={styles.leader_line_item}>{players[0]}, </span>
+        <span className={styles.leader_line_item}>{players[1]}</span>
+      </div>
+    </ErrorBoundary>
   );
 };
