+++ packages/client/src/styles/variables.scss
@@ -13,6 +13,7 @@ $progress_bar_bg_color: #222739;
 $progress_bar_border_color: black;
 
 $main_input_border_error: red;
+$boundary_error: red;
 
 $main_button_border_color: 3px solid rgb(0, 0, 0);
